FILTER UTILS: Pull simple_amb_measurement into filter utils
