+++ include/libswiftnav/filter_utils.h
@@ -17,6 +17,8 @@
 
 #include <common.h>
 
+double simple_amb_measurement(double carrier, double code);
+
 s8 assign_de_mtx(u8 num_sats, const sdiff_t *sats_with_ref_first,
                  const double ref_ecef[3], double *DE);
